refactor(db): split author_photos into cascade/remote/shared submodules
`db/src/author_photos/mod.rs` was 990 lines including a 479-line inline
test block. This commit completes the two-PR split described in #313:

PR 1 (tests): the inline `#[cfg(test)] mod tests` block was already moved
to `db/src/author_photos/tests.rs` in the branch.

PR 2 (prod split): split the 499-line production module by responsibility:
- `cascade.rs` — Open Library resolver (`resolve`, `resolve_with`,
  `OpenLibraryConfig`, `fetch_open_library`); `anyhow::Result` propagation
  preserved unchanged.
- `remote.rs` — SSRF-guarded remote image fetch (`fetch_remote_image`,
  `fetch_remote_image_with`, `FetchRemoteImageError`, `RemoteImageConfig`,
  `is_blocked_address`, `validated_resolve`).
- `shared.rs` — process-wide `reqwest::Client` singleton + `default_user_agent`.
- `mod.rs` — thin root that re-exports the public API and delegates to tests.

No public API changed. All 38 `author_photos::tests` pass unchanged.

Closes #313
