+++ CLAUDE.md
@@ -77,7 +77,8 @@ thumbs.rs           — F1.2 thumbnail pipeline: decode cover → resize to Sm/M
 indexer.rs          — scan → DB indexing, staleness checks (is_stale + reindex); reindex opts into materialize_sidecars. F2.3 `reindex_audiobooks` uses Phase A.5 group_into_books + Phase B parse_groups + sync_audiobooks (writes book_file_parts rows).
 library_layout.rs   — F0.6 canonical layout: slugify, canonical_path, sidecar_cover_for, allocate_canonical_path (collision suffix)
 worker/             — single-process Worker primitive: scan_sem (Scan/ScanAudiobooks) + hls_sem (HlsTranscode) semaphores + per-resource keyed mutex; Task::HlsTranscode dispatches into hls::transcode_book; WorkerConfig::hls_concurrency defaults to max(1, num_cpus/2). F1.11 ResolveAuthorPhoto dispatches into author_photos::resolve. Split into types (Task/Worker/config), queue (post/await_completion + RAII map-cleanup guards), exec (run dispatch loop), handlers (per-task-kind execute arms), progress (snapshot + retention + report_progress), tests
-author_photos.rs    — F1.11 author profile photo cascade (manual → Open Library → letter); injectable `OpenLibraryConfig` for tests (wiremock). First outbound HTTP in the codebase (reqwest, rustls-tls). Also hosts `fetch_remote_image` for the admin "paste URL" path — gated by a `RemoteImageConfig` SSRF guard (#275): resolves the host, blocks loopback / RFC1918 / link-local (incl. AWS IMDS 169.254.169.254) / multicast / IPv6 ULA / site-local / mapped-v4 before any TCP connect, then pins `reqwest` to the validated addresses so DNS rebinding can't slip in a private IP. The `allow_private_addresses` flag is test-only — production AppState construction always uses the strict default.
+author_photos/      — F1.11 author profile photo pipeline split into focused submodules: `cascade.rs` (manual → Open Library → letter resolution; `OpenLibraryConfig` for wiremock injection), `remote.rs` (`fetch_remote_image` + SSRF guard — `RemoteImageConfig`, `FetchRemoteImageError`; resolves host, blocks loopback/RFC1918/link-local/AWS IMDS/IPv6 ULA before any TCP connect, pins `reqwest` to validated addresses against DNS rebinding; `allow_private_addresses` is test-only), `shared.rs` (process-wide `reqwest::Client` singleton + `default_user_agent`), `tests.rs` (wiremock-backed cascade + SSRF guard unit tests). Public API re-exported from `mod.rs`.
+author_photos_data.rs — Row CRUD for the `author_photos` table (`get_author_photo`, `author_photo_status`, `upsert_author_photo`, `delete_author_photo`) plus `delete_author` admin primitive. `AuthorPhotoSource` enum. `AuthorPhotosDataError` wraps `sqlx::Error` per boundary rule.
 migrations/         — numbered SQL migrations embedded via sqlx::migrate!
 ```
