+++ db/src/author_photos/tests.rs
@@ -1,10 +1,19 @@
-use super::*;
-use crate::author_photos_data::get_author_photo;
-use crate::pool::init_db;
+use std::time::Duration;
+
+use sqlx::SqlitePool;
 use wiremock::{
     matchers::{method, path, query_param},
     Mock, MockServer, ResponseTemplate,
 };
+
+use super::cascade::{fetch_open_library, resolve_with, OpenLibraryConfig};
+use super::remote::{
+    fetch_remote_image_with, is_blocked_address, FetchRemoteImageError, RemoteImageConfig,
+};
+use crate::author_photos_data::{
+    author_photo_status, get_author_photo, upsert_author_photo, AuthorPhotoSource,
+};
+use crate::pool::init_db;
 
 async fn pool_with_author(name: &str) -> (SqlitePool, i64) {
     let pool = init_db("sqlite::memory:").await.unwrap();
